+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.0] - 2022-06-28
+
+- Added `FormBuilder` and `FieldGroupWrapperComponent`.
+
 ## [0.14.0] - 2022-06-28
 
-- Added FormBuilder.
+- Added `SideMenu::Component`.
 
 ## [0.13.0] - 2022-06-28